requirements.txt
absl-py==0.10.0
astunparse==1.6.3
autopep8==1.5.4
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
dlib==19.21.0
gast==0.3.3
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
isodate==0.6.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mpg123==0.4
msrest==0.6.18
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
toml==0.10.1
twilio==6.45.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
